Crate zng_ext_audio

Audio loading, rendering and cache.

Services

Services this extension provides.

AUDIOS

Crate

This crate is part of the zng project.

Cargo Features

This crate provides 1 feature flag, not enabled by default.

"http"

Enable web audio source.

Structs

AUDIOS

Audio loading, cache and render service.

AudioFormat

Represents an audio codec capability.

AudioFormatCapability

Capabilities of an AudioFormat implementation.

AudioHash

A 256-bit hash for audio tracks.

AudioHasher

Hasher that computes a AudioHash.

AudioLimits

Limits for audio loading and decoding.

AudioMix

Represents an audio source.

AudioOptions

Options for AUDIOS.audio.

AudioOutput

Represents an open audio output stream.

AudioOutputId

Unique identifier of an open audio output.

AudioOutputOptions

Options for a new audio output stream.

AudioTrack

State of an AudioVar.

AudioTracksMode

Defines what tracks are decoded from multi track containers.

WeakAudioOutput

Weak reference to an AudioOutput.

Enums

AudioCacheMode

Cache mode of AUDIOS.

AudioDataFormat

Format of the audio bytes.

AudioOutputState

Represents the playback state if an audio output stream.

AudioSource

The different sources of an audio resource.

AudioSourceFilter

Represents a PathFilter and UriFilter.

Traits

AudiosExtension

A custom extension for the AUDIOS service.

Type Aliases

AudioVar

Represents an AudioTrack tracked by the AUDIOS cache.

PathFilter

Represents an AudioSource::Read path request filter.

UriFilter

Represents an AudioSource::Download path request filter.